+++ protempa-framework/src/main/java/org/protempa/proposition/LocalUniqueIdValuesProvider.java
@@ -29,5 +29,5 @@ public interface LocalUniqueIdValuesProvider {
     
     String getId();
     
-    long getNumericalId();
+    int getNumericalId();
 }
